Oracle support in Cloud Controller

+++ spec/spec_helper.rb
@@ -52,6 +52,20 @@ module VCAP::CloudController
     def reset_database
       prepare_database
 
+      #Disable all constraints before dropping the tables
+      if db.database_type == :oracle
+        db.execute("BEGIN
+          FOR c IN
+          (SELECT c.owner, c.table_name, c.constraint_name
+           FROM user_constraints c, user_tables t
+           WHERE c.table_name = t.table_name
+           ORDER BY c.constraint_type DESC)
+          LOOP
+            dbms_utility.exec_ddl_statement('alter table \"' || c.owner || '\".\"' || c.table_name || '\" drop constraint ' || c.constraint_name);
+          END LOOP;
+        END;")
+      end
+      
       db.tables.each do |table|
         drop_table_unsafely(table)
       end
@@ -104,6 +118,10 @@ module VCAP::CloudController
           db.execute("SET foreign_key_checks = 0")
           db.drop_table(table)
           db.execute("SET foreign_key_checks = 1")
+
+        when :oracle
+          #Skip Oracle 'recycle bin' tables.  Let Oracle or a DBA take care of these.
+          db.drop_table(table) unless table.to_s.downcase.start_with?("bin$")
 
         # Postgres uses CASCADE directive in DROP TABLE
         # to remove foreign key contstraints.
